@@ -14,66 +14,67 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "4.3.5",
-    "@angular/cdk": "^2.0.0-beta.8",
-    "@angular/common": "4.3.5",
-    "@angular/compiler": "4.3.5",
-    "@angular/core": "4.3.5",
-    "@angular/forms": "4.3.5",
-    "@angular/http": "4.3.5",
-    "@angular/material": "^2.0.0-beta.11",
-    "@angular/platform-browser": "4.3.5",
-    "@angular/platform-browser-dynamic": "4.3.5",
-    "@angular/router": "4.3.5",
-    "@angular/service-worker": "^1.0.0-beta.16",
-    "@ngrx/core": "^1.0.0",
-    "@ngrx/effects": "^2.0.0",
-    "@ngrx/router-store": "^1.1.0",
-    "@ngrx/store": "^2.0.0",
-    "@ngrx/store-devtools": "^3.1.0",
-    "core-js": "^2.4.1",
-    "enhanced-resolve": "3.3.0",
+    "@angular/animations": "^6.0.3",
+    "@angular/cdk": "^6.1.0",
+    "@angular/common": "^6.0.3",
+    "@angular/compiler": "^6.0.3",
+    "@angular/core": "^6.0.3",
+    "@angular/forms": "^6.0.3",
+    "@angular/http": "^6.0.3",
+    "@angular/material": "^6.1.0",
+    "@angular/platform-browser": "^6.0.3",
+    "@angular/platform-browser-dynamic": "^6.0.3",
+    "@angular/router": "^6.0.3",
+    "@angular/service-worker": "^6.0.3",
+    "@ngrx/effects": "^6.0.1",
+    "@ngrx/router-store": "^6.0.1",
+    "@ngrx/store": "^6.0.1",
+    "@ngrx/store-devtools": "^6.0.1",
+    "core-js": "^2.5.4",
+    "enhanced-resolve": "4.0.0",
     "hammerjs": "2.0.8",
-    "intersection-observer": "^0.4.0",
-    "masonry-layout": "^4.2.0",
-    "moment": "^2.19.3",
-    "ng2-charts": "^1.5.0",
-    "reselect": "^2.5.4",
-    "rxjs": "^5.4.2",
+    "intersection-observer": "^0.5.0",
+    "jasmine": "^3.1.0",
+    "masonry-layout": "^4.2.1",
+    "moment": "^2.22.1",
+    "ng2-charts": "^1.6.0",
+    "reselect": "^3.0.1",
+    "rxjs": "^6.0.0",
     "screenfull": "^3.2.2",
-    "source-map": "^0.7.2",
-    "sw-precache": "^5.2.0",
+    "source-map": "^0.7.3",
+    "sw-precache": "^5.2.1",
     "sw-toolbox": "^3.6.0",
-    "ts-helpers": "^1.1.1",
-    "wicg-inert": "^1.1.4",
-    "zone.js": "^0.8.14"
+    "ts-helpers": "^1.1.2",
+    "wicg-inert": "^1.1.6",
+		"zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^1.1.0",
-    "@angular/compiler-cli": "4.3.5",
-    "@angular/language-service": "4.3.5",
-    "@types/jasmine": "~2.5.53",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "angular-cli-ghpages": "^0.5.0",
-    "codelyzer": "~3.1.1",
-    "jasmine-core": "~2.6.2",
-    "jasmine-marbles": "^0.1.0",
-    "jasmine-spec-reporter": "~4.1.0",
-    "karma": "~1.7.0",
-    "karma-chrome-launcher": "~2.1.1",
+    "@angular/cli": "~6.0.5",
+		"@angular/compiler-cli": "^6.0.3",
+		"@angular-devkit/build-angular": "~0.6.5",
+    "@angular/language-service": "^6.0.3",
+    "@types/jasmine": "~2.8.6",
+		"@types/jasminewd2": "~2.0.3",
+    "@types/node": "~8.9.4",
+    "angular-cli-ghpages": "^0.5.2",
+    "codelyzer": "~4.3.0",
+    "jasmine-core": "~2.99.1",
+    "jasmine-marbles": "^0.3.1",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~1.7.1",
+    "karma-chrome-launcher": "~2.2.0",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
+    "karma-coverage-istanbul-reporter": "~1.4.2",
+    "karma-jasmine": "~1.1.1",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "karma-remap-istanbul": "^0.2.2",
-    "ngrx-store-freeze": "^0.1.6",
-    "ngx-color-picker": "^4.1.0",
-    "protractor": "~5.1.2",
-    "sass-lint": "1.10.2",
-    "ts-node": "~3.2.0",
-    "tslint": "~5.3.2",
-    "typescript": "~2.3.4",
-    "webpack-bundle-analyzer": "^2.8.3"
+    "karma-remap-istanbul": "^0.6.0",
+    "ngrx-store-freeze": "^0.2.3",
+    "ngx-color-picker": "^6.3.3",
+    "protractor": "~5.3.0",
+    "sass-lint": "1.12.1",
+    "ts-node": "~5.0.1",
+    "tslint": "~5.9.1",
+		"typescript": "~2.7.2",
+		"webpack-bundle-analyzer": "^2.13.1"
   }
 }
