@@ -3,7 +3,7 @@
 import { TestBed, async } from '@angular/core/testing';
 import { ContactFormComponent } from './contact-form.component';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { HttpModule } from '@angular/http';
+import { HttpClientModule } from '@angular/common/http';
 
 const mockmessage = 'Lorem ipsum dolor sit amet, consectetur adipiscing elit.' +
 	'Maecenas pellentesque lorem eu nulla interdum, vel ornare quam sagittis.' +
@@ -17,7 +17,7 @@ describe('Component: ContactForm', () => {
 			imports: [
 				FormsModule,
 				ReactiveFormsModule,
-				HttpModule
+				HttpClientModule
 			],
 
 			declarations: [
