@@ -1,8 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Title } from '@angular/platform-browser';
-import { Observable } from 'rxjs/Observable';
-
-
+import { Observable } from 'rxjs';
 
 @Component({
 	selector: 'app-contact',
