@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { HttpModule } from '@angular/http';
+import { HttpClientModule } from '@angular/common/http';
 import { ReactiveFormsModule } from '@angular/forms';
 
 import { LoklakContactRoutingModule } from './contact-routing.module';
@@ -28,7 +28,7 @@ import { ContactFormComponent } from './contact-form/contact-form.component';
 		LoklakContactRoutingModule,
 		NavbarModule,
 		FooterModule,
-		HttpModule
+		HttpClientModule
 	],
 	declarations: [
 		ContactComponent,
