@@ -2,17 +2,17 @@ import { Injectable } from '@angular/core';
 import { Location } from '@angular/common';
 import { Title } from '@angular/platform-browser';
 import { Store, Action } from '@ngrx/store';
-import { Effect, Actions } from '@ngrx/effects';
-import { Observable } from 'rxjs/Observable';
-import { empty } from 'rxjs/observable/empty';
-import { of } from 'rxjs/observable/of';
-import 'rxjs/add/operator/catch';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/switchMap';
-import 'rxjs/add/operator/debounceTime';
-import 'rxjs/add/operator/skip';
-import 'rxjs/add/operator/takeUntil';
-import 'rxjs/add/operator/withLatestFrom';
+import { Effect, Actions, ofType } from '@ngrx/effects';
+import { Observable, of, empty } from 'rxjs';
+import {
+	catchError,
+	map,
+	switchMap,
+	withLatestFrom,
+	takeUntil,
+	debounceTime,
+	skip
+} from 'rxjs/operators';
 
 import { SearchService, SearchServiceConfig } from '../services';
 import * as fromRoot from '../reducers';
@@ -42,170 +42,205 @@ import { parseDateToApiAcceptedFormat } from '../utils';
 export class ApiSearchEffects {
 
 	@Effect()
-	search$: Observable<Action>
-		= this.actions$
-					.ofType(apiAction.ActionTypes.SEARCH)
-					.debounceTime(400)
-					.map((action: apiAction.SearchAction) => action.payload)
-					.switchMap(query => {
-						const searchServiceConfig: SearchServiceConfig = new SearchServiceConfig();
-						const nextSearch$ = this.actions$.ofType(apiAction.ActionTypes.SEARCH).skip(1);
+	search$: Observable<Action> = this.actions$
+		.pipe(
+			ofType(apiAction.ActionTypes.SEARCH),
+			debounceTime(400),
+			map((action: apiAction.SearchAction) => action.payload),
+			switchMap(query => {
+				const searchServiceConfig: SearchServiceConfig = new SearchServiceConfig();
+				const nextSearch$ = this.actions$
+					.pipe(
+						ofType(apiAction.ActionTypes.SEARCH),
+						skip(1)
+					);
 
-						searchServiceConfig.addAggregationFields(['created_at', 'screen_name', 'mentions', 'hashtags']);
-						if (query.filter.image) {
-							searchServiceConfig.addFilters(['image']);
-						} else {
-							searchServiceConfig.removeFilters(['image']);
-						}
-						if (query.filter.video) {
-							searchServiceConfig.addFilters(['video']);
-						} else {
-							searchServiceConfig.removeFilters(['video']);
-						}
+				searchServiceConfig.addAggregationFields(['created_at', 'screen_name', 'mentions', 'hashtags']);
+				if (query.filter.image) {
+					searchServiceConfig.addFilters(['image']);
+				} else {
+					searchServiceConfig.removeFilters(['image']);
+				}
+				if (query.filter.video) {
+					searchServiceConfig.addFilters(['video']);
+				} else {
+					searchServiceConfig.removeFilters(['video']);
+				}
 
-						return this.apiSearchService.fetchQuery(query.queryString, searchServiceConfig)
-												.takeUntil(nextSearch$)
-												.map(response => new apiAction.SearchCompleteSuccessAction(response))
-												.catch(() => of(new apiAction.SearchCompleteFailAction('')));
-					});
+				return this.apiSearchService
+					.fetchQuery(query.queryString, searchServiceConfig)
+					.pipe(
+						takeUntil(nextSearch$),
+						map(response => new apiAction.SearchCompleteSuccessAction(response)),
+						catchError(() => of(new apiAction.SearchCompleteFailAction('')))
+					);
+			})
+		);
 
 	@Effect()
-	trendingHashtagsSearch$: Observable<Action>
-		= this.actions$
-					.ofType(trendsAction.ActionTypes.SEARCH_TRENDING_HASHTAGS)
-					.debounceTime(400)
-					.map((action: trendsAction.SearchTrendingHashtagsAction) => action.payload)
-					.switchMap(_ => {
-						const searchServiceConfig: SearchServiceConfig = new SearchServiceConfig();
-						const nextRequest$ = this.actions$.ofType(trendsAction.ActionTypes.SEARCH_TRENDING_HASHTAGS).skip(1);
+	trendingHashtagsSearch$: Observable<Action> = this.actions$
+		.pipe(
+			ofType(trendsAction.ActionTypes.SEARCH_TRENDING_HASHTAGS),
+			debounceTime(400),
+			map((action: trendsAction.SearchTrendingHashtagsAction) => action.payload),
+			switchMap(_ => {
+				const searchServiceConfig: SearchServiceConfig = new SearchServiceConfig();
+				const nextRequest$ = this.actions$
+					.pipe(
+						ofType(trendsAction.ActionTypes.SEARCH_TRENDING_HASHTAGS),
+						skip(1)
+					);
 
-						searchServiceConfig.addAggregationFields(['hashtags']);
-						searchServiceConfig.count = 0;
-						searchServiceConfig.maximumRecords = 0;
-						searchServiceConfig.source = 'cache';
+				searchServiceConfig.addAggregationFields(['hashtags']);
+				searchServiceConfig.count = 0;
+				searchServiceConfig.maximumRecords = 0;
+				searchServiceConfig.source = 'cache';
 
-						const todayDate = new Date().toDateString();
-						const query = `since:${parseDateToApiAcceptedFormat(new Date(todayDate))}`;
+				const todayDate = new Date().toDateString();
+				const query = `since:${parseDateToApiAcceptedFormat(new Date(todayDate))}`;
 
-						return this.apiSearchService.fetchQuery(query, searchServiceConfig)
-												.takeUntil(nextRequest$)
-												.map(response => new apiAction.SearchTrendingHashtagsSuccessAction(response))
-												.catch(() => of(new apiAction.SearchTrendingHashtagsFailAction()));
-					});
+				return this.apiSearchService
+					.fetchQuery(query, searchServiceConfig)
+					.pipe(
+						takeUntil(nextRequest$),
+						map(response => new apiAction.SearchTrendingHashtagsSuccessAction(response)),
+						catchError(() => of(new apiAction.SearchTrendingHashtagsFailAction()))
+					);
+			})
+		);
 
 	@Effect()
 	relocateAfterSearchSuccess$: Observable<Action>
 		= this.actions$
-					.ofType(apiAction.ActionTypes.SEARCH_COMPLETE_SUCCESS,
-									apiAction.ActionTypes.SEARCH_COMPLETE_FAIL)
-					.withLatestFrom(this.store$)
-					.map(([action, state]) => {
-						return {
-							doRelocate: state.query.relocateAfter,
-							relocateTo: state.query.routerString
-						};
-					})
-					.map(relocateObject => {
-						if (relocateObject.doRelocate) {
-							const URIquery = encodeURIComponent(relocateObject.relocateTo);
+			.pipe(
+				ofType(
+					apiAction.ActionTypes.SEARCH_COMPLETE_SUCCESS,
+					apiAction.ActionTypes.SEARCH_COMPLETE_FAIL
+				),
+				withLatestFrom(this.store$),
+				map(([action, state]) => {
+					return {
+						doRelocate: state.query.relocateAfter,
+						relocateTo: state.query.routerString
+					};
+				}),
+				map(relocateObject => {
+					if (relocateObject.doRelocate) {
+						const URIquery = encodeURIComponent(relocateObject.relocateTo);
 
-							if (!this.location.isCurrentPathEqualTo(`/search?query=${URIquery}`)) {
-								this.location.go(`/search?query=${URIquery}`);
-							}
+						if (!this.location.isCurrentPathEqualTo(`/search?query=${URIquery}`)) {
+							this.location.go(`/search?query=${URIquery}`);
 						}
-						return new queryAction.RelocationAfterQueryResetAction();
-					});
+					}
+					return new queryAction.RelocationAfterQueryResetAction();
+				})
+			);
 
 	@Effect({ dispatch: false })
-	resetTitleAfterSearchSuccess$: Observable<void>
-		= this.actions$
-					.ofType(apiAction.ActionTypes.SEARCH_COMPLETE_SUCCESS,
-									apiAction.ActionTypes.SEARCH_COMPLETE_FAIL)
-					.withLatestFrom(this.store$)
-					.map(([action, state]) => {
-						const displayString = state.query.displayString;
-						let title = `${displayString} - Loklak Search`;
-						if (action.type === apiAction.ActionTypes.SEARCH_COMPLETE_FAIL) {
-							title += ' - No Results';
-						}
-						this.titleService.setTitle(title);
-					});
+	resetTitleAfterSearchSuccess$: Observable<void> = this.actions$
+		.pipe(
+			ofType(
+				apiAction.ActionTypes.SEARCH_COMPLETE_SUCCESS,
+				apiAction.ActionTypes.SEARCH_COMPLETE_FAIL
+			),
+			withLatestFrom(this.store$),
+			map(([action, state]) => {
+				const displayString = state.query.displayString;
+				let title = `${displayString} - Loklak Search`;
+				if (action.type === apiAction.ActionTypes.SEARCH_COMPLETE_FAIL) {
+					title += ' - No Results';
+				}
+				this.titleService.setTitle(title);
+			})
+		);
 
 	@Effect()
-	wallSearchAction$: Observable<Action>
-		= this.actions$
-					.ofType(wallAction.ActionTypes.WALL_SEARCH)
-					.debounceTime(400)
-					.map((action: wallAction.WallSearchAction) => action.payload)
-					.withLatestFrom(this.store$)
-					.map(([action, state]) => {
-						return {
-							directUrl: state.mediaWallDirectUrl.directUrl,
-							query: state.mediaWallQuery.query
-						};
-					})
-					.switchMap(queryObject => {
-						const nextSearch$ = this.actions$.ofType(wallAction.ActionTypes.WALL_SEARCH).skip(1);
-						const searchServiceConfig: SearchServiceConfig = new SearchServiceConfig();
+	wallSearchAction$: Observable<Action> = this.actions$
+		.pipe(
+			ofType(wallAction.ActionTypes.WALL_SEARCH),
+			debounceTime(400),
+			map((action: wallAction.WallSearchAction) => action.payload),
+			withLatestFrom(this.store$),
+			map(([action, state]) => {
+				return {
+					directUrl: state.mediaWallDirectUrl.directUrl,
+					query: state.mediaWallQuery.query
+				};
+			}),
+			switchMap(queryObject => {
+				const nextSearch$ = this.actions$
+					.pipe(
+						ofType(wallAction.ActionTypes.WALL_SEARCH),
+						skip(1)
+					);
+				const searchServiceConfig: SearchServiceConfig = new SearchServiceConfig();
 
-						if (queryObject.query.filter.image) {
-							searchServiceConfig.addFilters(['image']);
-						} else {
-							searchServiceConfig.removeFilters(['image']);
-						}
-						if (queryObject.query.filter.video) {
-							searchServiceConfig.addFilters(['video']);
-						} else {
-							searchServiceConfig.removeFilters(['video']);
-						}
+				if (queryObject.query.filter.image) {
+					searchServiceConfig.addFilters(['image']);
+				} else {
+					searchServiceConfig.removeFilters(['image']);
+				}
+				if (queryObject.query.filter.video) {
+					searchServiceConfig.addFilters(['video']);
+				} else {
+					searchServiceConfig.removeFilters(['video']);
+				}
 
-						searchServiceConfig.source = 'twitter';
+				searchServiceConfig.source = 'twitter';
 
-							return this.apiSearchService.fetchQuery(queryObject.query.queryString, searchServiceConfig)
-												.takeUntil(nextSearch$)
-												.map(response => {
-													const URIquery = encodeURIComponent(queryObject.query.routerString);
-													this.location.go(`/wall?${queryObject.directUrl}`);
-													return new apiAction.WallSearchCompleteSuccessAction(response);
-												})
-												.catch(() => of(new apiAction.WallSearchCompleteFailAction('')));
-					});
-
-	@Effect()
-	nextWallSearchAction$
-		= this.actions$
-					.ofType(apiAction.ActionTypes.WALL_SEARCH_COMPLETE_SUCCESS)
-					.debounceTime(10000)
-					.withLatestFrom(this.store$)
-					.map(([action, state]) => {
-							return new wallPaginationAction.WallNextPageAction('');
-					});
+					return this.apiSearchService
+						.fetchQuery(queryObject.query.queryString, searchServiceConfig)
+						.pipe(
+							takeUntil(nextSearch$),
+							map(response => {
+								const URIquery = encodeURIComponent(queryObject.query.routerString);
+								this.location.go(`/wall?${queryObject.directUrl}`);
+								return new apiAction.WallSearchCompleteSuccessAction(response);
+							}),
+							catchError(() => of(new apiAction.WallSearchCompleteFailAction('')))
+						);
+			})
+		);
 
 	@Effect()
-	nextWallSearchActionAfterFail$
-		= this.actions$
-					.ofType(apiAction.ActionTypes.WALL_SEARCH_COMPLETE_FAIL)
-					.debounceTime(5000)
-					.withLatestFrom(this.store$)
-					.map(([action, state]) => {
-							return new wallPaginationAction.WallNextPageAction('');
-					});
+	nextWallSearchAction$: Observable<Action> = this.actions$
+		.pipe(
+			ofType(apiAction.ActionTypes.WALL_SEARCH_COMPLETE_SUCCESS),
+			debounceTime(10000),
+			withLatestFrom(this.store$),
+			map(([action, state]) => {
+					return new wallPaginationAction.WallNextPageAction('');
+			})
+		);
+
+	@Effect()
+	nextWallSearchActionAfterFail$: Observable<Action> = this.actions$
+		.pipe(
+			ofType(apiAction.ActionTypes.WALL_SEARCH_COMPLETE_FAIL),
+				debounceTime(5000),
+				withLatestFrom(this.store$),
+				map(([action, state]) => {
+						return new wallPaginationAction.WallNextPageAction('');
+				})
+		);
 
 	@Effect({ dispatch: false })
-	resetTitleAfterWallSearchSuccess$: Observable<void>
-		= this.actions$
-					.ofType(apiAction.ActionTypes.WALL_SEARCH_COMPLETE_SUCCESS,
-									apiAction.ActionTypes.WALL_SEARCH_COMPLETE_FAIL)
-					.withLatestFrom(this.store$)
-					.map(([action, state]) => {
-						const displayString = state.mediaWallQuery.query.displayString;
-						let title = `${displayString} - Loklak Media Wall`;
-						if (action.type === apiAction.ActionTypes.WALL_SEARCH_COMPLETE_FAIL) {
-							title += ' - No Results';
-						}
-						this.titleService.setTitle(title);
-					});
-
+	resetTitleAfterWallSearchSuccess$: Observable<void> = this.actions$
+		.pipe(
+			ofType(
+				apiAction.ActionTypes.WALL_SEARCH_COMPLETE_SUCCESS,
+				apiAction.ActionTypes.WALL_SEARCH_COMPLETE_FAIL
+			),
+			withLatestFrom(this.store$),
+			map(([action, state]) => {
+				const displayString = state.mediaWallQuery.query.displayString;
+				let title = `${displayString} - Loklak Media Wall`;
+				if (action.type === apiAction.ActionTypes.WALL_SEARCH_COMPLETE_FAIL) {
+					title += ' - No Results';
+				}
+				this.titleService.setTitle(title);
+			})
+		);
 
 	constructor(
 		private store$: Store<fromRoot.State>,
