@@ -1,15 +1,8 @@
 import { Injectable } from '@angular/core';
-import { Effect, Actions } from '@ngrx/effects';
+import { Effect, Actions, ofType } from '@ngrx/effects';
 import { Action } from '@ngrx/store';
-import { Observable } from 'rxjs/Observable';
-import { empty } from 'rxjs/observable/empty';
-import { of } from 'rxjs/observable/of';
-import 'rxjs/add/operator/catch';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/switchMap';
-import 'rxjs/add/operator/debounceTime';
-import 'rxjs/add/operator/skip';
-import 'rxjs/add/operator/takeUntil';
+import { Observable, of, empty } from 'rxjs';
+import { map, switchMap, debounceTime, takeUntil, skip, catchError } from 'rxjs/operators';
 
 import { UserService } from '../services';
 import * as userApiAction from '../actions/user-api';
@@ -31,18 +24,27 @@ import { Query } from '../models';
 export class ApiUserSearchEffects {
 
 	@Effect()
-	search$: Observable<Action>
-		= this.actions$
-					.ofType(userApiAction.ActionTypes.USER_SEARCH)
-					.debounceTime(400)
-					.map((action: userApiAction.UserSearchAction) => action.payload)
-					.switchMap(query => {
-						const nextSearch$ = this.actions$.ofType(userApiAction.ActionTypes.USER_SEARCH).skip(1);
-						return this.apiUserService.fetchQuery(query.screen_name)
-												.takeUntil(nextSearch$)
-												.map(response => new userApiAction.UserSearchCompleteSuccessAction(response))
-												.catch(() => of(new userApiAction.UserSearchCompleteFailAction('')));
-					});
+	search$: Observable<Action> = this.actions$
+		.pipe(
+			ofType(userApiAction.ActionTypes.USER_SEARCH),
+			debounceTime(400),
+			map((action: userApiAction.UserSearchAction) => action.payload),
+			switchMap(query => {
+				const nextSearch$ = this.actions$.pipe(
+					ofType(userApiAction.ActionTypes.USER_SEARCH),
+					skip(1)
+				);
+
+				const follow_count = 10;
+
+				return this.apiUserService.fetchQuery(query.screen_name)
+					.pipe(
+						takeUntil(nextSearch$),
+						map(response => new userApiAction.UserSearchCompleteSuccessAction(response)),
+						catchError(() => of(new userApiAction.UserSearchCompleteFailAction('')))
+					);
+			})
+		);
 
 	constructor(
 		private actions$: Actions,
