@@ -1,11 +1,9 @@
 import { Injectable } from '@angular/core';
 import { Location } from '@angular/common';
 import { Store, Action } from '@ngrx/store';
-import { Effect, Actions } from '@ngrx/effects';
-import { Observable } from 'rxjs/Observable';
-import { of } from 'rxjs/observable/of';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/withLatestFrom';
+import { Effect, Actions, ofType } from '@ngrx/effects';
+import { Observable, of } from 'rxjs';
+import { map, withLatestFrom } from 'rxjs/operators';
 
 import * as fromRoot from '../reducers';
 import * as mediaWallDirectUrlAction from '../actions/media-wall-direct-url';
@@ -32,46 +30,48 @@ export class MediaWallDirectUrlEffects {
 	@Effect()
 	generateDirectUrl$: Observable<Action>
 		= this.actions$
-		.ofType(mediaWallDirectUrlAction.ActionTypes.WALL_GENERATE_DIRECT_URL)
-		.withLatestFrom(this.store$)
-		.map(([action, state]) => {
-			return {
-				query: state.mediaWallQuery.query,
-				design: state.mediaWallDesign.design,
-				hiddenFeedId: state.mediaWallResponse.response.hiddenFeedId,
-				blockedUser: state.mediaWallResponse.response.blockedUser,
-				profanityCheck: state.mediaWallResponse.response.profanityCheck,
-				removeDuplicate: state.mediaWallResponse.response.removeDuplicate,
-				wallHeader: state.mediaWallCustom.wallHeader,
-				wallCard: state.mediaWallCustom.wallCard,
-				wallBackground: state.mediaWallCustom.wallBackground
-			};
-		})
-		.map(queryObject => {
-			const configSet = {
-				query: queryObject.query.displayString,
-				imageFilter: queryObject.query.filter.image,
-				location: queryObject.query.location,
-				displayHeader: queryObject.design.displayHeader,
-				columnCount: queryObject.design.columnCount,
-				count: queryObject.design.count,
-				cardStyle: queryObject.design.cardStyle,
-				headerTitle: queryObject.design.headerTitle,
-				hiddenFeedId: queryObject.hiddenFeedId,
-				blockedUser: queryObject.blockedUser,
-				profanityCheck: queryObject.profanityCheck,
-				removeDuplicate: queryObject.removeDuplicate,
-				wallHeaderBackgroundColor: queryObject.wallHeader.backgroundColor,
-				wallHeaderFontColor: queryObject.wallHeader.fontColor,
-				wallCardFontColor: queryObject.wallCard.fontColor,
-				wallCardAccentColor: queryObject.wallCard.accentColor,
-				wallCardBackgroundColor: queryObject.wallCard.backgroundColor,
-				wallBackgroundColor: queryObject.wallBackground.backgroundColor
-			}
-			const directUrl = generateDirectUrl(configSet);
-			this.location.go(`/wall?${directUrl}`);
-			return new mediaWallDirectUrlAction.WallShortenDirectUrlAction(directUrl);
-			});
+		.pipe(
+			ofType(mediaWallDirectUrlAction.ActionTypes.WALL_GENERATE_DIRECT_URL),
+			withLatestFrom(this.store$),
+			map(([action, state]) => {
+				return {
+					query: state.mediaWallQuery.query,
+					design: state.mediaWallDesign.design,
+					hiddenFeedId: state.mediaWallResponse.response.hiddenFeedId,
+					blockedUser: state.mediaWallResponse.response.blockedUser,
+					profanityCheck: state.mediaWallResponse.response.profanityCheck,
+					removeDuplicate: state.mediaWallResponse.response.removeDuplicate,
+					wallHeader: state.mediaWallCustom.wallHeader,
+					wallCard: state.mediaWallCustom.wallCard,
+					wallBackground: state.mediaWallCustom.wallBackground
+				};
+			}),
+			map(queryObject => {
+				const configSet = {
+					query: queryObject.query.displayString,
+					imageFilter: queryObject.query.filter.image,
+					location: queryObject.query.location,
+					displayHeader: queryObject.design.displayHeader,
+					columnCount: queryObject.design.columnCount,
+					count: queryObject.design.count,
+					cardStyle: queryObject.design.cardStyle,
+					headerTitle: queryObject.design.headerTitle,
+					hiddenFeedId: queryObject.hiddenFeedId,
+					blockedUser: queryObject.blockedUser,
+					profanityCheck: queryObject.profanityCheck,
+					removeDuplicate: queryObject.removeDuplicate,
+					wallHeaderBackgroundColor: queryObject.wallHeader.backgroundColor,
+					wallHeaderFontColor: queryObject.wallHeader.fontColor,
+					wallCardFontColor: queryObject.wallCard.fontColor,
+					wallCardAccentColor: queryObject.wallCard.accentColor,
+					wallCardBackgroundColor: queryObject.wallCard.backgroundColor,
+					wallBackgroundColor: queryObject.wallBackground.backgroundColor
+				};
+				const directUrl = generateDirectUrl(configSet);
+				this.location.go(`/wall?${directUrl}`);
+				return new mediaWallDirectUrlAction.WallShortenDirectUrlAction(directUrl);
+			})
+		);
 
 	constructor(
 		private actions$: Actions,
