@@ -1,14 +1,9 @@
 import { Injectable } from '@angular/core';
 import { Location } from '@angular/common';
-import { Effect, Actions } from '@ngrx/effects';
+import { Effect, Actions, ofType } from '@ngrx/effects';
 import { Store, Action } from '@ngrx/store';
-import { Observable } from 'rxjs/Observable';
-import { of } from 'rxjs/observable/of';
-import 'rxjs/add/operator/catch';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/switchMap';
-import 'rxjs/add/operator/takeUntil';
-import 'rxjs/add/operator/withLatestFrom';
+import { Observable, of } from 'rxjs';
+import { catchError, map, switchMap, takeUntil, withLatestFrom, debounceTime } from 'rxjs/operators';
 
 import { SearchService, SearchServiceConfig } from '../services';
 import { Query } from '../models';
@@ -36,61 +31,67 @@ export class WallPaginationEffects {
 	private searchServiceConfig: SearchServiceConfig = new SearchServiceConfig();
 
 	@Effect()
-	wallPagination$: Observable<Action>
-		= this.actions$
-					.ofType(wallPaginationAction.ActionTypes.WALL_NEXT_PAGE)
-					.map((action: wallPaginationAction.WallNextPageAction) => action.payload)
-					.withLatestFrom(this.store$)
-					.map(([action, state]) => {
-						return {
-							query: state.mediaWallQuery.query,
-							count: state.mediaWallDesign.design.count
-						};
-					})
-					.switchMap(queryObject => {
-						const nextSearch$ = this.actions$.ofType(wallAction.ActionTypes.WALL_SEARCH);
+	wallPagination$: Observable<Action> = this.actions$
+		.pipe(
+			ofType(wallPaginationAction.ActionTypes.WALL_NEXT_PAGE),
+			map((action: wallPaginationAction.WallNextPageAction) => action.payload),
+			withLatestFrom(this.store$),
+			map(([action, state]) => {
+				return {
+					query: state.mediaWallQuery.query,
+					count: state.mediaWallDesign.design.count
+				};
+			}),
+			switchMap(queryObject => {
+				const nextSearch$ = this.actions$.ofType(wallAction.ActionTypes.WALL_SEARCH);
 
-						if (queryObject.query.filter.image) {
-							this.searchServiceConfig.addFilters(['image']);
-						} else {
-							this.searchServiceConfig.removeFilters(['image']);
-						}
-						if (queryObject.query.filter.video) {
-							this.searchServiceConfig.addFilters(['video']);
-						} else {
-							this.searchServiceConfig.removeFilters(['video']);
-						}
-						this.searchServiceConfig.source = 'twitter';
-						this.searchServiceConfig.count = queryObject.count;
-						this.searchServiceConfig.maximumRecords = queryObject.count;
+				if (queryObject.query.filter.image) {
+					this.searchServiceConfig.addFilters(['image']);
+				} else {
+					this.searchServiceConfig.removeFilters(['image']);
+				}
+				if (queryObject.query.filter.video) {
+					this.searchServiceConfig.addFilters(['video']);
+				} else {
+					this.searchServiceConfig.removeFilters(['video']);
+				}
+				this.searchServiceConfig.source = 'twitter';
+				this.searchServiceConfig.count = queryObject.count;
+				this.searchServiceConfig.maximumRecords = queryObject.count;
 
-						return this.apiSearchService.fetchQuery(queryObject.query.queryString, this.searchServiceConfig)
-												.takeUntil(nextSearch$)
-												.map(response => {
-													return new wallPaginationAction.WallPaginationCompleteSuccessAction(response);
-												})
-												.catch(() => of(new wallPaginationAction.WallPaginationCompleteFailAction('')));
-					});
-
-	@Effect()
-	nextWallSearchAction$
-		= this.actions$
-					.ofType(wallPaginationAction.ActionTypes.WALL_PAGINATION_COMPLETE_SUCCESS)
-					.debounceTime(10000)
-					.withLatestFrom(this.store$)
-					.map(([action, state]) => {
-						return new wallPaginationAction.WallNextPageAction('');
-					});
+				return this.apiSearchService
+					.fetchQuery(queryObject.query.queryString, this.searchServiceConfig)
+					.pipe(
+						takeUntil(nextSearch$),
+						map(response => {
+							return new wallPaginationAction.WallPaginationCompleteSuccessAction(response);
+						}),
+						catchError(() => of(new wallPaginationAction.WallPaginationCompleteFailAction('')))
+					);
+			})
+		);
 
 	@Effect()
-		nextWallSearchActionAfterFail$
-		= this.actions$
-					.ofType(wallPaginationAction.ActionTypes.WALL_PAGINATION_COMPLETE_FAIL)
-					.debounceTime(5000)
-					.withLatestFrom(this.store$)
-					.map(([action, state]) => {
-						return new wallPaginationAction.WallNextPageAction('');
-					});
+	nextWallSearchAction$ = this.actions$
+		.pipe(
+			ofType(wallPaginationAction.ActionTypes.WALL_PAGINATION_COMPLETE_SUCCESS),
+			debounceTime(10000),
+			withLatestFrom(this.store$),
+			map(([action, state]) => {
+				return new wallPaginationAction.WallNextPageAction('');
+			})
+		);
+
+	@Effect()
+		nextWallSearchActionAfterFail$ = this.actions$
+			.pipe(
+				ofType(wallPaginationAction.ActionTypes.WALL_PAGINATION_COMPLETE_FAIL),
+				debounceTime(5000),
+				withLatestFrom(this.store$),
+				map(([action, state]) => {
+					return new wallPaginationAction.WallNextPageAction('');
+				})
+			);
 
 	constructor(
 		private actions$: Actions,
