@@ -1,14 +1,9 @@
 import { Injectable } from '@angular/core';
 import { Location } from '@angular/common';
-import { Effect, Actions } from '@ngrx/effects';
+import { Effect, Actions, ofType } from '@ngrx/effects';
 import { Store, Action } from '@ngrx/store';
-import { Observable } from 'rxjs/Observable';
-import { of } from 'rxjs/observable/of';
-import 'rxjs/add/operator/catch';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/operator/switchMap';
-import 'rxjs/add/operator/takeUntil';
-import 'rxjs/add/operator/withLatestFrom';
+import { Observable, of } from 'rxjs';
+import { map, switchMap, takeUntil, withLatestFrom, catchError } from 'rxjs/operators';
 
 import { SearchService, SearchServiceConfig } from '../services';
 import { Query } from '../models';
@@ -35,40 +30,49 @@ export class PaginationEffects {
 	private searchServiceConfig: SearchServiceConfig = new SearchServiceConfig();
 
 	@Effect()
-	pagination$: Observable<Action>
-		= this.actions$
-					.ofType(paginationAction.ActionTypes.NEXT_PAGE)
-					.map((action: paginationAction.NextPageAction) => action.payload)
-					.withLatestFrom(this.store$)
-					.map(([action, state]) => {
-						return {
-							query: state.query,
-							lastRecord: state.apiResponse.entities.length,
-						};
-					})
-					.switchMap(queryObject => {
-						const nextSearch$ = this.actions$.ofType(apiAction.ActionTypes.SEARCH);
+	pagination$: Observable<Action> = this.actions$
+		.pipe(
+			ofType(paginationAction.ActionTypes.NEXT_PAGE),
+			map((action: paginationAction.NextPageAction) => action.payload),
+			withLatestFrom(this.store$),
+			map(([action, state]) => {
+				return {
+					query: state.query,
+					lastRecord: state.apiResponse.entities.length,
+				};
+			}),
+			switchMap(queryObject => {
+				const nextSearch$ = this.actions$.ofType(apiAction.ActionTypes.SEARCH);
 
-						this.searchServiceConfig.startRecord = queryObject.lastRecord + 1;
-						this.searchServiceConfig.addAggregationFields(['created_at', 'screen_name', 'mentions', 'hashtags']);
-						if (queryObject.query.filter.image) {
-							this.searchServiceConfig.addFilters(['image']);
-						} else {
-							this.searchServiceConfig.removeFilters(['image']);
-						}
-						if (queryObject.query.filter.video) {
-							this.searchServiceConfig.addFilters(['video']);
-						} else {
-							this.searchServiceConfig.removeFilters(['video']);
-						}
+				this.searchServiceConfig.startRecord = queryObject.lastRecord + 1;
+				this.searchServiceConfig.addAggregationFields([
+					'created_at',
+					'screen_name',
+					'mentions',
+					'hashtags'
+				]);
+				if (queryObject.query.filter.image) {
+					this.searchServiceConfig.addFilters(['image']);
+				} else {
+					this.searchServiceConfig.removeFilters(['image']);
+				}
+				if (queryObject.query.filter.video) {
+					this.searchServiceConfig.addFilters(['video']);
+				} else {
+					this.searchServiceConfig.removeFilters(['video']);
+				}
 
-						return this.apiSearchService.fetchQuery(queryObject.query.queryString, this.searchServiceConfig)
-												.takeUntil(nextSearch$)
-												.map(response => {
-													return new paginationAction.PaginationCompleteSuccessAction(response);
-												})
-												.catch(() => of(new paginationAction.PaginationCompleteFailAction('')));
-					});
+				return this.apiSearchService
+					.fetchQuery(queryObject.query.queryString, this.searchServiceConfig)
+					.pipe(
+						takeUntil(nextSearch$),
+						map(response => {
+							return new paginationAction.PaginationCompleteSuccessAction(response);
+						}),
+						catchError(() => of(new paginationAction.PaginationCompleteFailAction('')))
+					);
+			})
+		);
 
 	constructor(
 		private actions$: Actions,
