@@ -8,7 +8,7 @@ import {
 	ChangeDetectionStrategy,
 	ElementRef
 } from '@angular/core';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 
 import { Store } from '@ngrx/store';
 import * as fromRoot from '../../reducers';
@@ -77,22 +77,19 @@ export class FeedAdvancedSearchComponent implements OnInit {
 				image: false,
 				video: false
 			};
-		}
-		else if (value === 'image') {
+		} else if (value === 'image') {
 			this.selectedTab = 'image';
 			this.filterList = {
 				image: true,
 				video: false
 			};
-		}
-		else if (value === 'video') {
+		} else if (value === 'video') {
 			this.selectedTab = 'video';
 			this.filterList = {
 				image: false,
 				video: true
 			};
-		}
-		else {
+		} else {
 			this.selectedTab = 'all';
 			this.filterList = {
 				image: false,
@@ -110,24 +107,21 @@ export class FeedAdvancedSearchComponent implements OnInit {
 				since: null,
 				until: null
 			};
-		}
-		else if (value === 'lastDay') {
+		} else if (value === 'lastDay') {
 			this.timeBoundValue = 'Past 24 hours';
 			const date24HoursBefore = new Date(new Date().getTime() - (24 * 60 * 60 * 1000));
 			this.timeBound = {
 				since: date24HoursBefore,
 				until: null
 			};
-		}
-		else if (value === 'lastWeek') {
+		} else if (value === 'lastWeek') {
 			this.timeBoundValue = 'Past week';
 			const date1WeekBefore = new Date(new Date().getTime() - (7 * 24 * 60 * 60 * 1000));
 			this.timeBound = {
 				since: date1WeekBefore,
 				until: null
 			};
-		}
-		else {
+		} else {
 			this.timeBoundValue = 'Any time';
 			this.timeBound = {
 				since: null,
@@ -142,24 +136,19 @@ export class FeedAdvancedSearchComponent implements OnInit {
 		if (value === 'all') {
 			this.locationValue = 'All Countries';
 			this.location = null;
-		}
-		else if (value === 'India') {
+		} else if (value === 'India') {
 			this.locationValue = 'Country: India';
 			this.location = 'India';
-		}
-		else if (value === 'China') {
+		} else if (value === 'China') {
 			this.locationValue = 'Country: China';
 			this.location = 'China';
-		}
-		else if (value === 'US') {
+		} else if (value === 'US') {
 			this.locationValue = 'Country: US';
 			this.location = 'US';
-		}
-		else if (value === 'UK') {
+		} else if (value === 'UK') {
 			this.locationValue = 'Country: UK';
 			this.location = 'UK';
-		}
-		else {
+		} else {
 			this.locationValue = 'All Countries';
 			this.location = null;
 		}
