@@ -1,10 +1,7 @@
 import { Injectable } from '@angular/core';
-import { Jsonp, Response, URLSearchParams } from '@angular/http';
-import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/operator/catch';
-import 'rxjs/add/operator/map';
-import 'rxjs/add/observable/throw';
-import 'rxjs/add/operator/retry';
+import { HttpClient, HttpParams } from '@angular/common/http';
+import { Observable, throwError } from 'rxjs';
+import { catchError, map, retry } from 'rxjs/operators';
 
 import { UserResponse } from '../models/api-user-response';
 
@@ -16,30 +13,27 @@ export class UserService {
 	private static following = '4';
 
 	constructor(
-		private jsonp: Jsonp
+		private http: HttpClient
 	) { }
 
 	// TODO: make the searchParams as configureable model rather than this approach.
 	public fetchQuery(user: string): Observable<UserResponse> {
 		const screen_name = user.charAt(0).toUpperCase() + user.slice(1);
-		const searchParams = new URLSearchParams();
+		const searchParams = new HttpParams();
 		searchParams.set('screen_name', screen_name);
 		searchParams.set('followers', UserService.followers);
 		searchParams.set('following', UserService.following);
 		searchParams.set('callback', 'JSONP_CALLBACK');
 		searchParams.set('minified', UserService.minified_results.toString());
-		return this.jsonp.get(UserService.apiUrl.toString(), {search : searchParams})
-								.map(this.extractData)
-								.retry(2)
-								.catch(this.handleError);
-	}
 
-	private extractData(res: Response): UserResponse {
-		try {
-			return <UserResponse>res.json();
-		} catch (error) {
-			console.error(error);
-		}
+		const jsonpUrl = UserService.apiUrl.toString() + `?${searchParams.toString()}`;
+
+		return this.http
+			.jsonp<UserResponse>(jsonpUrl, 'callback')
+			.pipe(
+				retry(2),
+				catchError(this.handleError)
+			);
 	}
 
 	private handleError (error: any) {
@@ -47,6 +41,6 @@ export class UserService {
 		const errMsg = (error.message) ? error.message :
 									error.status ? `${error.status} - ${error.statusText}` : 'Server error';
 		console.error(errMsg); // Right now we are logging to console itself
-		return Observable.throw(errMsg);
+		return throwError(errMsg);
 	}
 }
