@@ -5,7 +5,12 @@ import { RouterTestingModule } from '@angular/router/testing';
 import { BrowserDynamicTestingModule } from '@angular/platform-browser-dynamic/testing';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { TestBed, async } from '@angular/core/testing';
-import { MatDialogModule, MatDialog } from '@angular/material';
+import {
+	MatDialogModule,
+	MatDialog,
+	MatIconModule,
+	MatTabsModule
+} from '@angular/material';
 import { MediaWallCustomizationComponent } from './media-wall-customization.component';
 import { StoreModule } from '@ngrx/store';
 import * as fromRoot from '../../reducers';
@@ -44,7 +49,9 @@ describe('MediaWallCustomizationComponent', () => {
 				RouterTestingModule,
 				BrowserAnimationsModule,
 				MatDialogModule,
-				StoreModule.provideStore(fromRoot.reducer),
+				MatIconModule,
+				MatTabsModule,
+				StoreModule.forRoot(fromRoot.reducers),
 				FormsModule,
 				ReactiveFormsModule
 			],
