@@ -8,7 +8,9 @@ import {
 	MatDialogModule,
 	MatDialog,
 	MatTooltipModule,
-	MatMenuModule
+	MatMenuModule,
+	MatIconModule,
+	MatGridListModule
 } from '@angular/material';
 import { StoreModule } from '@ngrx/store';
 import * as fromRoot from '../../reducers';
@@ -26,8 +28,10 @@ describe('MediaWallMenuComponent', () => {
 				BrowserAnimationsModule,
 				MatDialogModule,
 				MatMenuModule,
+				MatIconModule,
+				MatGridListModule,
 				MatTooltipModule,
-				StoreModule.provideStore(fromRoot.reducer),
+				StoreModule.forRoot(fromRoot.reducers),
 				FormsModule,
 				ReactiveFormsModule
 			],
