@@ -11,7 +11,9 @@ import {
 	MatDialog,
 	MatSlideToggleModule,
 	MatSelectModule,
-	MatCheckboxModule
+	MatCheckboxModule,
+	MatIconModule,
+	MatTabsModule
 } from '@angular/material';
 import { RouterTestingModule } from '@angular/router/testing';
 import { StoreModule } from '@ngrx/store';
@@ -20,7 +22,7 @@ import { MediaWallQueryComponent } from './media-wall-query/media-wall-query.com
 import { MediaWallModerationComponent } from './media-wall-moderation/media-wall-moderation.component';
 import { MediaWallDesignComponent } from './media-wall-design/media-wall-design.component';
 
-import { reducer } from '../reducers';
+import { reducers } from '../reducers';
 import { RouterStub } from '../../testing';
 import { MediaWallComponent } from './media-wall.component';
 import { MasonryModule } from '../app-masonry/app-masonry.module';
@@ -49,7 +51,7 @@ class MediaWallLinkerStubComponent {
 	@Input() unshorten;
 	@Input() useAll;
 	@Input() wallCustomText;
-	@Output() onShowed;
+	@Output() showed;
 }
 
 @Component({
@@ -59,7 +61,7 @@ class MediaWallLinkerStubComponent {
 class MediaWallHeaderStubComponent {
 	@Input() query;
 	@Input() wallCustomHeader;
-	@Input() headerTitle
+	@Input() headerTitle;
 }
 
 @Component({
@@ -114,9 +116,11 @@ describe('Component: MediaWall', () => {
 				MatSlideToggleModule,
 				MatSelectModule,
 				MatCheckboxModule,
+				MatTabsModule,
+				MatIconModule,
 				MasonryModule,
 				LazyImgModule,
-				StoreModule.provideStore(reducer)
+				StoreModule.forRoot(reducers)
 			],
 			declarations: [
 				MediaWallComponent,
