@@ -1,34 +1,20 @@
-import 'zone.js/dist/long-stack-trace-zone';
-import 'zone.js/dist/proxy.js';
-import 'zone.js/dist/sync-test';
-import 'zone.js/dist/jasmine-patch';
-import 'zone.js/dist/async-test';
-import 'zone.js/dist/fake-async-test';
+// This file is required by karma.conf.js and loads recursively all the .spec and framework files
 
-import 'hammerjs';
+import 'zone.js/dist/zone-testing';
+import { getTestBed } from '@angular/core/testing';
+import {
+	BrowserDynamicTestingModule,
+	platformBrowserDynamicTesting
+} from '@angular/platform-browser-dynamic/testing';
 
-// Unfortunately there's no typing for the `__karma__` variable. Just declare it as any.
-declare var __karma__: any;
-declare var require: any;
+declare const require: any;
 
-// Prevent Karma from running prematurely.
-__karma__.loaded = function () { };
-
-
-Promise.all([
-	System.import('@angular/core/testing'),
-	System.import('@angular/platform-browser-dynamic/testing')
-])
-	// First, initialize the Angular testing environment.
-	.then(([testing, testingBrowser]) => {
-		testing.getTestBed().initTestEnvironment(
-			testingBrowser.BrowserDynamicTestingModule,
-			testingBrowser.platformBrowserDynamicTesting()
-		);
-	})
-	// Then we find all the tests.
-	.then(() => require.context('./', true, /\.spec\.ts/))
-	// And load the modules.
-	.then(context => context.keys().map(context))
-	// Finally, start Karma to run the tests.
-	.then(__karma__.start, __karma__.error);
+// First, initialize the Angular testing environment.
+getTestBed().initTestEnvironment(
+	BrowserDynamicTestingModule,
+	platformBrowserDynamicTesting()
+);
+// Then we find all the tests.
+const context = require.context('./', true, /\.spec\.ts$/);
+// And load the modules.
+context.keys().map(context);
